minor changes in anticipation of filter functionality (#808)
* minor changes in anticipation of filter functionality

* adds new Filters.svelte container
